* Moved VehicleStopMode to enums package * Moved toInt functionality of VehicleStopMode to SUMO package and changed around the method signatures a little bit for more cleanliness
